pyproject.toml
[project]
name = "automcp"
version = "0.1.0"
description = ""
authors = [
    {name = "0range"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "mcp (>=1.26.0,<2.0.0)",
    "langchain (>=1.2.8,<2.0.0)",
    "langgraph (>=1.0.7,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sse-starlette (>=3.2.0,<4.0.0)",
    "langchain-openai (>=1.1.9,<2.0.0)",
    "onnxruntime (==1.23.2)",
    "chromadb (>=1.5.0,<2.0.0)",
    "sentence-transformers (>=5.2.2,<6.0.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "ollama (>=0.6.1,<0.7.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "psutil (>=5.9.0,<6.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "langchain-anthropic (>=1.3.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
asyncio_mode = "auto"